RDKBDEV-1068 rbus: find gcno files in current directory
Reason for change: In dunfell, the gcno files are created in
the build directory itself (with the full path as the filename, slashes
replaced by hashes) rather than in the src and unittest directories.
Thus, the find command that creates the gcno tarball doesn't find
anything.
Change the find command to use the current directory instead of the src
and unittests sub-directories. This works both in case the gcno files
are created in the current directory, and in case they are created in
subdirectories.
Risks: None

Signed-off-by: Mohamed Shaikh <mohamed.shaikh@t-systems.com>
Change-Id: I10b5cced2dad39600e2c7455061987aa661ff4b5

Change-Id: I4a7e8412534bd62c2fa8b770e9d8b49ebdc6a7a5