RDKCOM-3205: RDKBDEV-1068 rbus: find gcno files in current directory
Reason for change: In dunfell, the gcno files are created in
the build directory itself (with the full path as the filename, slashes
replaced by hashes) rather than in the src and unittest directories.
Thus, the find command that creates the gcno tarball doesn't find
anything.
Change the find command to use the current directory instead of the src
and unittests sub-directories. This works both in case the gcno files
are created in the current directory, and in case they are created in
subdirectories.
Risks: None
Signed-off-by: Mohamed Shaikh <mohamed.shaikh@t-systems.com>
Test Procedure: None.

(cherry picked from commit 3e1970bb8b4f8324b7dda10e9ee83b41692955a8)
Change-Id: Ib5ab6d35b2ac224ee0cde6bb4fc2bdaa13bea3af